model: Define class 'TrackItemView'
In our database, item categories are maintained in a
separate table. The category 'symbol' isn't directly
accessible to the adapter classes.

This class represents 'resolved' view of the item.
As of now, this means that category-id is resolved
to corresponding symbols.

This object shall be initialized via SQL 'join' query.
Hence, only getters are implemented.

Signed-off-by: Sanjeev Premi <spremi@ymail.com>